#108 Refactor / LightSBToDevice 1. Light StructuredBuffer 생성 Device로 옮김 2. CRenderMgr_Init에서 LightSB를 생성하진 않지만 멤버로 두고 CDevice Get함수를 통해 설정해주는 것이 좋은가 아님 그냥 매번 UpdateData에서 Get함수로 가져오는 것이 좋은가
