+++ Project/Engine/CRenderMgr_Init.cpp
@@ -13,12 +13,6 @@
 
 void CRenderMgr::init()
 {
-	m_Light2DBuffer = new CStructuredBuffer;
-	m_Light2DBuffer->Create(sizeof(tLightInfo), 10, SB_TYPE::READ_ONLY, true);
-
-	m_Light3DBuffer = new CStructuredBuffer;
-	m_Light3DBuffer->Create(sizeof(tLightInfo), 10, SB_TYPE::READ_ONLY, true);
-
 	m_pDebugObj = new CGameObject;
 	m_pDebugObj->AddComponent(new CTransform);
 	m_pDebugObj->AddComponent(new CMeshRender);
